+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unrelease
 
+## [0.6.3] - 2021-07-01
+### Added
+- ability to generate Stacks Addresses
+
+### Changed
+- deprecated Blockstack file extensions, refactored to extensions package
+
 ## [0.6.2] - 2020-11-19
 ### Added
 - ability to decrypt using the EncryptedResult and a BigInteger Private Key